% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{d12_to_dproj}
\alias{d12_to_dproj}
\title{convert d12 = d1+d2 to projected dose sqrt(d1^2+d2^2) in the direction of (d1, d2)
 given fixed ratio d2/d1}
\usage{
d12_to_dproj(d12, d2.d1)
}
\arguments{
\item{d12}{d12 = d1+d2 to projected dose sqrt(d1^2+d2^2) in the direction of (d1, d2).}

\item{d2.d1}{fixed ratio d2/d1.}
}
\description{
convert d12 = d1+d2 to projected dose sqrt(d1^2+d2^2) in the direction of (d1, d2)
 given fixed ratio d2/d1
}
