% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{d1_to_dproj}
\alias{d1_to_dproj}
\title{convert d1 to projected dose sqrt(d1^2+d2^2) given fixed ratio d2/d1}
\usage{
d1_to_dproj(d1, d2.d1)
}
\arguments{
\item{d1}{d1 to projected dose sqrt(d1^2+d2^2).}

\item{d2.d1}{fixed ratio d2/d1.}
}
\description{
convert d1 to projected dose sqrt(d1^2+d2^2) given fixed ratio d2/d1
}
