% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{d2_to_dproj}
\alias{d2_to_dproj}
\title{convert d2 to projected dose sqrt(d1^2+d2^2) given fixed ratio d2/d1}
\usage{
d2_to_dproj(d2, d2.d1, islogd = FALSE)
}
\arguments{
\item{d2}{d1 to projected dose sqrt(d1^2+d2^2).}

\item{d2.d1}{fixed ratio d2/d1.}

\item{islogd}{loged or not. Default is set to FLASE.}
}
\description{
convert d2 to projected dose sqrt(d1^2+d2^2) given fixed ratio d2/d1
}
