% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_extradata.R
\docType{data}
\name{datNCI60}
\alias{datNCI60}
\title{Example raw data for calculating GI50, TGI and LC50}
\format{An object of class \code{data.frame} with 12 rows and 2 columns.}
\usage{
datNCI60
}
\description{
This dataset is an example data set where the response is between -1 and 1.
 It is a 2 column data frame where the Dose column is the concentration of drug 
 and Growth column is growth percentage as defined by the NCI60 method.
}
\references{
Holbeck, S. L., Collins, J. M., & Doroshow, J. H. (2010). Analysis of food
 and drug administration approved anticancer agents in the NCI60 panel of human tumor cell lines.
 Molecular cancer therapeutics, 9(5), 1451-1460.
}
\keyword{datasets}
