% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{dproj_to_d2}
\alias{dproj_to_d2}
\title{convert projected dose sqrt(d1^2+d2^2) to d2 given fixed ratio d2/d1}
\usage{
dproj_to_d2(dproj, d2.d1, islogd = FALSE)
}
\arguments{
\item{dproj}{projected dose sqrt(d1^2+d2^2) to d2.}

\item{d2.d1}{fixed ratio d2/d1.}

\item{islogd}{loged or not. Default is set to FLASE.}
}
\description{
convert projected dose sqrt(d1^2+d2^2) to d2 given fixed ratio d2/d1
}
