% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{drOutlier}
\alias{drOutlier}
\title{Identifying outliers for dose-response data}
\usage{
drOutlier(drMat, alpha = 0.01)
}
\arguments{
\item{drMat}{dose-response matrix. The first column being dosage and the second column
being response. controls are included by specifying dose = 0.}

\item{alpha}{a scalar for significance level. Only 0.01, 0.05 and 1 are allowed. 
alpha = 1 is included for comparability issue. In this case, no outliers will be identified.}
}
\value{
indicator a vector with length nrow(drMat) specifying if a data point is outlier. 
 Control points always have status FALSE
}
\description{
Identifying outliers for dose-response data
}
\examples{
data(ryegrass, package = 'drc') # use the ryegrass data from drc package
drOutlier(drMat = ryegrass[, c(2, 1)], alpha = 0.05)
}
