% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{fitHSA_ray_boot}
\alias{fitHSA_ray_boot}
\title{bootstrap of fitHSA_ray to obtain variation of deltaAUC}
\usage{
fitHSA_ray_boot(
  d1,
  d2,
  e,
  tol = 0.05,
  strict = TRUE,
  name1 = "Drug A",
  name2 = "Drug B",
  dmin = NULL,
  dmax = NULL,
  islogd = TRUE,
  bootstrapB = 10,
  nrep = 3,
  seed = 100
)
}
\arguments{
\item{d1}{drug 1 dose in original scale.}

\item{d2}{drug 2 dose in original scale.}

\item{e}{response at the dose pair.}

\item{tol}{tolerance in declaring fixed ratio.}

\item{strict}{whether to stop if not fixed ratio; this is just for visualization;
the calculations of synergy is wrong since d12 = d1+d2 is meaningless for variable ratio! 
At the end, this is a waste of time: it would not make sense and impossible to do this!}

\item{name1}{drug 1 name.}

\item{name2}{drug 2 name.}

\item{dmin}{start position to calcualte AUC at projected dose, in original scale,
default is minimum projected dose.}

\item{dmax}{end position to calcualte AUC at projected dose, in original scale,
default is maximum projected dose.}

\item{islogd}{whether to compute AUC at log10 dose; if TRUE, dmin and dmax must also be
log10 scale.}

\item{bootstrapB}{number of boostrap iterations.}

\item{nrep}{how many replicates are sampled with replacement for each dose combination.
nrep=3 will make it much narrower CI band.}

\item{seed}{random seed. Dafault is set to 100.}
}
\description{
bootstrap of fitHSA_ray to obtain variation of deltaAUC
}
