% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_IAI.ModelSelection.R
\name{fitIAI_mod0}
\alias{fitIAI_mod0}
\title{Re-implemented IAI calculation with model selection and CI}
\usage{
fitIAI_mod0(d1, d2, e, E = seq(0.05, 0.95, 0.005), name1 = "Drug A",
  name2 = "Drug B", ratio)
}
\arguments{
\item{d1}{dose for drug 1.}

\item{d2}{dose for drug 2.}

\item{e}{corresponding percentile in the range [0, 1]; response is 1-E.}

\item{E}{a vector of responses (between 0 and 1) where IAI and confidence interval are to be
computed from.}

\item{name1}{drug 1 name.}

\item{name2}{drug 2 name.}

\item{ratio}{ratio of d1/d2.}
}
\value{
a list.
}
\description{
Re-implemented IAI calculation with model selection and CI
}
