% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getNormPvec}
\alias{getNormPvec}
\title{calculate P value assuming avg ~ Normal(mean, sd)}
\usage{
getNormPvec(avg, mean, sd)
}
\arguments{
\item{avg}{average.}

\item{mean}{mean.}

\item{sd}{standard deviation.}
}
\description{
calculate P value assuming avg ~ Normal(mean, sd)
}
