% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hill.R, R/drexplorer_nci60.R
\name{lines.hillFit}
\alias{lines.hillFit}
\alias{lines,hillFit}
\alias{lines.nci60Fit}
\alias{lines,nci60Fit}
\title{lines method for hillFit class}
\usage{
\method{lines}{hillFit}(
  fit,
  col = 1,
  lwd = 2,
  show_points = FALSE,
  pcol = "black",
  pch = 16,
  ...
)

\method{lines}{nci60Fit}(
  fit,
  pcol = "black",
  lcol = "black",
  lwd = 2,
  lty = 1,
  pch = 16,
  type = c("line")
)
}
\arguments{
\item{fit}{a hillFit object as returned by hillFit.}

\item{col}{line color.}

\item{lwd}{line width.}

\item{show_points}{whether to add points for dose-response paires (dose not 0).}

\item{pcol}{color for points; only relevant if show_points is TRUE.}

\item{pch}{pch for points; only relevant if show_points is TRUE.}

\item{...}{additional parametrs passed to generic lines() function.}

\item{lcol}{line color.}

\item{lty}{line type.}

\item{type}{default is set to line.}
}
\description{
lines method for hillFit class

lines method for nci60Fit class
}
