% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hill.R, R/drexplorer_nci60.R
\name{plot.hillFit}
\alias{plot.hillFit}
\alias{plot,hillFit}
\alias{plot.nci60Fit}
\alias{plot,nci60Fit}
\title{plot method for hillFit class}
\usage{
\method{plot}{hillFit}(
  fit,
  xlab = "Log10(Dose)",
  ylab = "Relative viability",
  xlim = NULL,
  ylim = NULL,
  main = "Fit of Hill equation",
  cex.main = 1,
  cex.axis = 1,
  pcol = "black",
  lcol = "black",
  lwd = 2,
  ...
)

\method{plot}{nci60Fit}(
  fit,
  xlab = "Dose",
  ylab = "Relative growth",
  main = "Fit of NCI60 method",
  xlim = NULL,
  ylim = NULL,
  cex.main = 1,
  cex.axis = 1,
  cex.lab = 1,
  axes = TRUE,
  pcol = "black",
  pch = 1,
  lcol = "black",
  lty = 1,
  ltyh = 1,
  lwd = 2,
  type = c("line", "points"),
  h = c(-0.5, 0, 0.5),
  ...
)
}
\arguments{
\item{fit}{a hillFit object as returned by hillFit.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{xlim}{y limit for display.}

\item{ylim}{x limit for display.}

\item{main}{main title.}

\item{cex.main}{cex for main title.}

\item{cex.axis}{cex for axis annotation.}

\item{pcol}{color for points.}

\item{lcol}{color for lines.}

\item{lwd}{line width used to specify the appearance of fitted curve.}

\item{...}{additional parameters, not implemented.}

\item{cex.lab}{cex for label.}

\item{axes}{axes.}

\item{pch}{point size.}

\item{lty}{line type.}

\item{ltyh}{horizontal line type.}

\item{type}{default is set to lines and points.}

\item{h}{horizontal line to add indicating e.g. GI (h=0.5), TGI (h=0), LD50 (h=-0.5)}
}
\description{
plot method for hillFit class

plot method for nci60Fit class
}
