% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_HSA.Bliss.R
\name{plotHSA_ray}
\alias{plotHSA_ray}
\title{plot the ray design from fit object returned by fitHSA_ray}
\usage{
plotHSA_ray(fit, cols = c(1, 2, 3), lwds = 3, xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = "Relative viability",
  main = NULL, tag = NULL, type = c("line", "HSA", "Bliss", "se"),
  legend = TRUE, legend.cex = 0.6, plotSetup = TRUE, digit = 3)
}
\arguments{
\item{fit}{the fitted result from fitHSA_ray().}

\item{cols}{color corresponding to mixture, drugA, drugB.}

\item{lwds}{line width.}

\item{xlim}{x-axis limit.}

\item{ylim}{y-axis limit.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{figure title.}

\item{tag}{title tag.}

\item{type}{plot types.}

\item{legend}{whether to show legend.}

\item{legend.cex}{cex of the legend.}

\item{plotSetup}{whether initiate a new figure; set to False when just to add curve as lines.}

\item{digit}{Default is set to 3.}
}
\description{
plot the ray design from fit object returned by fitHSA_ray
}
