% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_IAI.ModelSelection.R
\name{plotIAI_mod}
\alias{plotIAI_mod}
\title{Plot for the fitted object from model selection}
\usage{
plotIAI_mod(fitL, plotFA = F, mixtureDose = c("DualAxisA"), ylim = c(0, 1.2))
}
\arguments{
\item{fitL}{the fitted result from fitIAI_mod().}

\item{plotFA}{whether to plot FA value; if FALSE, dose will be used as x-axis for
IAI estimates}

\item{mixtureDose}{options to plot mixture dose. This can be (1) 'A+B':  mixture dose is
the sum of the 2 drug doses, (2) 'A': mixture dose represented by drug 1 dose, 
(3) 'B': mixture dose represented by drug 2 dose, (4) 'DualAxisA' where dual axis is used
for drug 1 and drug 2 dose but mixture represented by drug 1. This is useful if the two drugs
have very different doses, (5) 'DualAxisB', similar to 'DualAxisA' but mixture is represented
by drug 2 dose.}

\item{ylim}{y axis limits. Default is set to 0 to 1.2.}
}
\description{
Plot for the fitted object from model selection
}
