% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,drFit-method}
\title{predict method for drFit object}
\usage{
predict(object, ...)

\S4method{predict}{drFit}(object, newData)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{...}{additional arguments affecting the predictions produced.}

\item{newData}{a vector of dose levels to predict at. Default is to predict response at observed dose levels.}
}
\value{
a vector of predicted values
}
\description{
predict method for drFit object
}
