% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_hill.R, R/drexplorer_nci60.R
\name{predict.hillFit}
\alias{predict.hillFit}
\alias{predict,hillFit}
\alias{predict.nci60Fit}
\alias{predict,nci60Fit}
\title{predict method for hillFit class}
\usage{
\method{predict}{hillFit}(fit, newData = NULL)

\method{predict}{nci60Fit}(fit, newData = NULL)
}
\arguments{
\item{fit}{a hillFit/nci60Fit object as returned by hillFit/nci60Fit.}

\item{newData}{dose in original scale to be used for prediction of response.}
}
\value{
predicted response at specified dose.
}
\description{
predict method for hillFit class

predict method for nci60Fit class
}
