% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer_wrapper.R
\name{prepDRdat2expInfo}
\alias{prepDRdat2expInfo}
\title{From prepared data to experiment setup information}
\usage{
prepDRdat2expInfo(resPrepDR)
}
\arguments{
\item{resPrepDR}{result returned by prepDRdat}
}
\value{
a data frame
}
\description{
This includes: N dose (non-control), N technical rep (number of obs. at each dose,
 may be a string), experimental variation including medianSD, minSD, maxSD, meanSD
 computed from SD across different tech replicates.
}
