% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexplorer.R
\name{recommendedModels}
\alias{recommendedModels}
\title{Frequently used models in drexplorer2}
\usage{
recommendedModels()
}
\value{
a vector of recommended models
}
\description{
This function returns frequently used dose-response models for anti-cancer drug screening.
}
