% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_extradata.R
\docType{data}
\name{ryegrass}
\alias{ryegrass}
\title{ryegrass dataset from drc package}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
ryegrass
}
\description{
This dataset is from the ryegrass data in drc package. It is a 2 column data frame where the
 conc column is the concentration of ferulic acid in mM and rootl column is the root length of
 ryegrass.
}
\references{
Inderjit and J. C. Streibig, and M. Olofsdotter (2002) Joint action of phenolic
 acid mixtures and its significance in allelopathy research, Physiologia Plantarum, 114,
 422-428, 2002.
}
\keyword{datasets}
